requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bilibili-api-python==16.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.1
MarkupSafe==2.1.5
multidict==6.0.5
pillow==10.2.0
pyasn1==0.6.0
pycryptodomex==3.20.0
pypng==0.20220715.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
qrcode-terminal==0.8
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
